telegram_bot/bot/requirements.txt
pyTelegramBotAPI==3.6.6
psycopg2==2.7.7
python-dotenv==0.10.1